Move some methods into EnrollmentStepDefinitions
They are not just limited to validation or data coverage tests.

Issue #518 Implement accessibility checking in CI
