+++ psm-app/integration-tests/src/test/java/gov/medicaid/features/enrollment/steps/ValidationStepDefinitions.java
@@ -1,6 +1,5 @@
 package gov.medicaid.features.enrollment.steps;
 
-import cucumber.api.java.en.Given;
 import cucumber.api.java.en.Then;
 import cucumber.api.java.en.When;
 import gov.medicaid.features.general.steps.GeneralSteps;
@@ -23,23 +22,6 @@ public class ValidationStepDefinitions {
 
     private OrganizationInfoPage organizationInfoPage;
     private PersonalInfoPage personalInfoPage;
-
-    @Given("^I have started an enrollment$")
-    public void i_have_started_an_enrollment() {
-        generalSteps.loginAsProvider();
-        enrollmentSteps.createEnrollment();
-    }
-
-    @Given("^I am on the organization page$")
-    public void i_am_on_the_organization_page() {
-        enrollmentSteps.selectOrganizationalProviderType();
-    }
-
-    @Given("^I am on the personal info page$")
-    public void i_am_on_the_personal_info_page() {
-        i_have_started_an_enrollment();
-        enrollmentSteps.selectIndividualProviderType();
-    }
 
     @When("^when I enter an (\\d+) digit FEIN$")
     public void when_I_enter_an_digit_FEIN(int numDigits) {
